refactor: 예약 생성 관련 검증 로직 분리
