+++ src/main/java/roomescape/reservation/service/ReservationFactoryService.java
@@ -1,7 +1,6 @@
 package roomescape.reservation.service;
 
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
 import roomescape.member.domain.Member;
 import roomescape.member.repository.MemberRepository;
 import roomescape.reservation.domain.Reservation;
@@ -13,7 +12,6 @@ import roomescape.reservation.repository.ReservationTimeRepository;
 import roomescape.reservation.repository.ThemeRepository;
 
 @Service
-@Transactional(readOnly = true)
 public class ReservationFactoryService {
 
     private final ReservationTimeRepository reservationTimeRepository;
